performance-service/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
supabase==2.28.3
pydantic-settings==2.6.1
PyJWT==2.12.0
slowapi==0.1.9
pyodbc==5.2.0
tenacity==9.1.2
cachetools==5.5.2
httpx==0.27.2
apscheduler==3.10.4
